+++ core/src/test/java/com/opensymphony/xwork2/config/providers/ConfigurationProviderOgnlAllowlistTest.java
@@ -50,6 +50,7 @@ public class ConfigurationProviderOgnlAllowlistTest extends XWorkJUnit4TestCase 
                 Class.forName("java.io.Serializable"),
                 Class.forName("com.opensymphony.xwork2.mock.MockResult"),
                 Class.forName("com.opensymphony.xwork2.interceptor.ConditionalInterceptor"),
+                Class.forName("org.apache.struts2.ActionSupport"),
                 Class.forName("com.opensymphony.xwork2.ActionSupport"),
                 Class.forName("com.opensymphony.xwork2.ActionChainResult"),
                 Class.forName("com.opensymphony.xwork2.TextProvider"),
@@ -82,6 +83,7 @@ public class ConfigurationProviderOgnlAllowlistTest extends XWorkJUnit4TestCase 
                 Class.forName("java.io.Serializable"),
                 Class.forName("com.opensymphony.xwork2.mock.MockResult"),
                 Class.forName("com.opensymphony.xwork2.interceptor.ConditionalInterceptor"),
+                Class.forName("org.apache.struts2.ActionSupport"),
                 Class.forName("com.opensymphony.xwork2.ActionSupport"),
                 Class.forName("com.opensymphony.xwork2.TextProvider"),
                 Class.forName("com.opensymphony.xwork2.interceptor.Interceptor"),
@@ -111,6 +113,7 @@ public class ConfigurationProviderOgnlAllowlistTest extends XWorkJUnit4TestCase 
                 Class.forName("com.opensymphony.xwork2.LocaleProvider"),
                 Class.forName("java.io.Serializable"),
                 Class.forName("com.opensymphony.xwork2.interceptor.ConditionalInterceptor"),
+                Class.forName("org.apache.struts2.ActionSupport"),
                 Class.forName("com.opensymphony.xwork2.ActionSupport"),
                 Class.forName("com.opensymphony.xwork2.ActionChainResult"),
                 Class.forName("com.opensymphony.xwork2.TextProvider"),
